sandbox-apps/swagbot/src/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
bytecode==0.15.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
ddtrace==2.15.0
Deprecated==1.2.14
distro==1.9.0
envier==0.5.2
Flask==3.0.3
Flask-Cors==5.0.0
flask-openai==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.45.1
opentelemetry-api==1.27.0
protobuf==5.28.2
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
xmltodict==0.13.0
zipp==3.20.2
vertexai
python-json-logger==2.0.7